move tests for moved classes
